pyproject.toml
[project]
name = "cloneacademy"
version = "0.1.0"
description = "AI-powered GitHub repository documentation generator"
authors = [
    {name = "Ashish", email = "ashish.gec@gmail.com"}
]
dependencies = [
    "streamlit>=1.32.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "gitpython>=3.1.0",
    "PyGithub>=2.2.0",
    "Jinja2>=3.1.0",
    "typing-extensions>=4.5.0",
    "urllib3>=2.0.0",
    "certifi>=2023.7.22",
    "charset-normalizer>=3.2.0",
    "idna>=3.4",
    "PyYAML>=6.0.1",
    "python-dateutil>=2.8.2",
    "six>=1.16.0",
    "MarkupSafe>=2.1.3",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
] 